pyproject.toml
[tool.poetry]
name = "day_statistic_backend"
version = "0.1.0"
description = ""
authors = ["Levantin Oleg <levantin.oleg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
pytest = "^7.1.2"
email-validator = "^1.2.1"
python-dotenv = "^0.21.0"
beanie = "^1.11.9"
redis-om = "^0.0.27"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
autoflake = "^1.4"
isort = "^5.10.1"
black = "^22.6.0"
flake8 = "^5.0.4"
pyright = "^1.1.262"
ansible = "^6.2.0"
httpx = "^0.23.0"
pytest-anyio = "^0.0.0"
asgi-lifespan = "^1.0.1"

[tool.autoflake]
remove-all-unused-imports = true
ignore-init-module-imports = true
remove-unused-variables = true
in-place = true
recursive = true

[tool.pyright]
reportMissingParameterType = true
reportUnknownParameterType = true

[tool.isort]
profile = "black"
line_length = 79

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
